backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.17
langchain-openai==0.1.20
langchain-community==0.2.17
# faiss-cpu==1.8.0
tiktoken
python-multipart
pypdf2
pydantic-settings